Add cadastro de agentes, fix login

+++ code/backend/src/main/kotlin/example/micronaut/Auth/Agente/AgenteService.kt
@@ -3,15 +3,16 @@ package example.micronaut.autentificacao.agente
 import example.micronaut.autentificacao.usuario.User
 import example.micronaut.autentificacao.usuario.UserRepository
 import jakarta.inject.Singleton
+import jakarta.transaction.Transactional
 
 @Singleton
-class AgenteService(
+open class AgenteService(
     private val agenteRepository: AgenteRepository,
     private val userRepository: UserRepository
 ) {
 
-    fun criar(dto: AgenteDTO): Agente {
-
+    @Transactional
+    open fun criar(dto: AgenteDTO): Agente {
         if (userRepository.existsByEmail(dto.email)) {
             throw RuntimeException("Email já cadastrado")
         }
@@ -20,7 +21,7 @@ class AgenteService(
             throw RuntimeException("CNPJ já cadastrado")
         }
 
-        val savedUser = userRepository.save(
+        val userSalvo = userRepository.save(
             User(
                 email = dto.email,
                 password = dto.password
@@ -28,10 +29,11 @@ class AgenteService(
         )
 
         val agente = Agente(
+            id = userSalvo.id,
             tipo = dto.tipo,
             cnpj = dto.cnpj,
             razaoSocial = dto.razaoSocial,
-            user = savedUser
+            user = userSalvo
         )
 
         return agenteRepository.save(agente)
@@ -51,30 +53,31 @@ class AgenteService(
             .orElseThrow { RuntimeException("Agente não encontrado") }
     }
 
-    fun atualizar(id: Long, dto: AgenteDTO): Agente {
+    @Transactional
+    open fun atualizar(id: Long, dto: AgenteDTO): Agente {
         val agenteExistente = buscarPorId(id)
-
         val userExistente = agenteExistente.user
 
         userExistente.email = dto.email
         userExistente.password = dto.password
 
-        val savedUser = userRepository.update(userExistente)
+        userRepository.update(userExistente)
 
         agenteExistente.tipo = dto.tipo
         agenteExistente.cnpj = dto.cnpj
         agenteExistente.razaoSocial = dto.razaoSocial
-        agenteExistente.user = savedUser
 
         return agenteRepository.update(agenteExistente)
     }
 
-    fun deletar(id: Long) {
+    @Transactional
+    open fun deletar(id: Long) {
         val agente = buscarPorId(id)
+        val userId = agente.user.id
 
         agenteRepository.deleteById(id)
 
-        agente.user.id?.let { userId ->
+        if (userId != null) {
             userRepository.deleteById(userId)
         }
     }
